backend/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Pillow==10.4.0
Werkzeug==2.3.7
python-dotenv==1.0.0
email-validator==2.0.0
twilio==8.9.1
email-validator==2.0.0